Add infrastructure for role-based AI
This commit adds the basic instructions (for a human) and infrastructure
for use with a ChatGPT project to instruct the AI assistant to have
swappable roles.

I chose this approach over using custom GPTs because they don't
currently support using a custom GPT within a project, and projects are
the only organizational structure that OpenAI supports.

At the moment none of the features unique to custom GPTs are more
compelling than having a single folder per project.